changing import approach to be more robust
