+++ slurm_scripts/run_layer_norm.sh
@@ -7,7 +7,7 @@
 #SBATCH --gres=gpu:1          
 #SBATCH --cpus-per-task=16
 #SBATCH --time=01:00:00  
-#SBATCH --mem=150GB
+#SBATCH --mem=100GB
 #SBATCH --partition=kempner_requeue
 
 # Source common functions
@@ -28,22 +28,23 @@ BATCH_SIZE=${8:-256}
 DOMAIN_CONFIG=${9:-"domains.ini"}
 DOMAIN_ID=${10:-"B"}
 USE_STANDARD_DATASET=${11:-"false"}
-COMPARISON_DIR=${12:-""}
-DEBUG_MODE=${13:-"false"}
+DEBUG_MODE=${12:-"false"}
 
 export DOMAIN_ID=$DOMAIN_ID
 export DOMAIN_CONFIG=$DOMAIN_CONFIG
-export EXPERIMENT_TYPE="comparison"
-export COMPARISON_DIR=$COMPARISON_DIR
+export EXPERIMENT_TYPE=${EXPERIMENT_TYPE:-"layer_norm"}  # Use inherited value or default to basic
 export DEBUG_MODE=$DEBUG_MODE
+export PYTHONPATH="/n/home00/cberon/code/Transformers_for_Modeling_Decision_Sequences/utils:$PYTHONPATH"
+export PYTHONPATH="/n/home00/cberon/code/Transformers_for_Modeling_Decision_Sequences/transformer:$PYTHONPATH"
+export PYTHONPATH="/n/home00/cberon/code/Transformers_for_Modeling_Decision_Sequences:$PYTHONPATH"
+
 
 # Setup standard dataset if requested
 setup_standard_dataset \
     --use-standard-dataset "$USE_STANDARD_DATASET" \
-    --comparison-dir "$COMPARISON_DIR" \
     --domain-config "$DOMAIN_CONFIG" \
     --domain-id "$DOMAIN_ID" \
-    --multiple-domains "true" \
+    --multiple-domains "false" \
     --train-steps "$TRAIN_STEPS" \
     --val-steps "1000000" \
     --run-number "$RUN_NUMBER"
@@ -63,8 +64,7 @@ else
         --num_steps_train=$TRAIN_STEPS \
         --num_steps_val=1_000_000 \
         --no_overwrite \
-        --config_file "$DOMAIN_CONFIG" \
-        --multiple_domains
+        --config_file "$DOMAIN_CONFIG"
     
     # Run evaluation for individual datasets
     python ${BASE_PATH}/evaluation/basic_evaluation.py --run $RUN_NUMBER
@@ -83,7 +83,8 @@ srun --cpu-bind=none python -m transformer.train \
     --n_head=$N_HEAD \
     --n_embd=$EMBD_DIM \
     --sequence_length=$CONTEXT_LENGTH \
-    --checkpoint_interval="log"
+    --checkpoint_interval="log" \
+    --model_type="GPT_noLN"
 
 # Setup GPU environment for inference
 setup_gpu_environment
@@ -94,3 +95,26 @@ python ${INFERENCE_PATH}/guess_using_transformer.py --run $RUN_NUMBER
 python ${INFERENCE_PATH}/evaluate_transformer_guess.py --run $RUN_NUMBER
 python ${INFERENCE_PATH}/graphs_transformer_vs_ground_truth.py --run $RUN_NUMBER
 python ${BASE_PATH}/evaluation/inspect_data.py --run $RUN_NUMBER
+
+# # Process checkpoints in parallel
+# print_section_header "Processing Checkpoints"
+# process_checkpoints
+
+# # Plot checkpoint comparison
+# print_section_header "Checkpoint Comparison"
+# python ${INFERENCE_PATH}/plot_checkpoint_comparison.py --run $RUN_NUMBER
+
+# print_section_header "Learning Analysis"
+# LEARNING_COMMANDS=(
+#     "python ${INFERENCE_PATH}/learning.py --run $RUN_NUMBER --step_max=100"
+#     "python ${INFERENCE_PATH}/learning.py --run $RUN_NUMBER --step_max=1000"
+#     "python ${INFERENCE_PATH}/learning.py --run $RUN_NUMBER --step_min=1000 --step_max=10000"
+#     "python ${INFERENCE_PATH}/learning.py --run $RUN_NUMBER --step_min=10000 --step_max=100000"
+#     "python ${INFERENCE_PATH}/learning.py --run $RUN_NUMBER"
+# )
+
+# # Run learning commands (will run sequentially if only 1 GPU)
+# run_on_gpus "${LEARNING_COMMANDS[@]}"
+
+print_section_header "Interpretability"
+python ${BASE_PATH}/interpretability/interp_sampler.py --run $RUN_NUMBER
