+++ transformer/transformer.py
@@ -8,10 +8,10 @@ import torch.distributed as dist
 import torch.nn as nn
 from torch.nn import functional as F
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+# sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from utils.file_management import get_experiment_file, read_sequence
 
-# Set random seed for reproducibility
+
 seed = 200
 torch.manual_seed(seed)
 if torch.cuda.is_available():
